VSC_RDS_Pre/requirements.txt
numpy
pandas
mysql
